+++ packages/component/src/Styles/defaultStyleSetOptions.js
@@ -40,6 +40,10 @@ const DEFAULT_OPTIONS = {
   bubbleMinWidth: 250, // min screen width = 300px, Edge requires 372px (https://developer.microsoft.com/en-us/microsoft-edge/platform/issues/13621468/)
   bubbleTextColor: 'Black',
 
+  // Root
+  rootHeight: '100%',
+  rootWidth: '100%',
+  
   // Send box
   hideSendBox: false,
   hideUploadButton: false,